Cleaned up scanner functions and strategies (extracted common functions and removed checkName), removed save to database and scanner dockercompose
