--- scanners/docker-compose-scanners.yaml
@@ -1,60 +1,0 @@
-# Separate docker compose for just the scanners. 
-version: '1'
-services:
-
-  endpoint-dispatcher:
-    build:
-      context: ./endpoint-dispatcher
-    network_mode: host
-    
-  github-clone-repo:
-    build:
-      context: ./github-clone-repo
-    network_mode: host
-
-  # github-cloned-dockerignore-check:
-  #   build:
-  #     context: ./github-cloned-dockerignore-check
-  #   # ports:
-  #   #   - 4002:4000 
-
-  github-cloned-gitignore-check:
-    build:
-      context: ./github-cloned-gitignore-check
-    network_mode: host
-
-  github-cloned-has-dependabot-yaml-check:
-    build:
-      context: ./github-cloned-has-dependabot-yaml-check
-    network_mode: host
-
-  github-cloned-has-security-md-check:
-    build:
-      context: ./github-cloned-has-security-md-check
-    network_mode: host
-
-  github-cloned-has-tests-directory-check:
-    build:
-      context: ./github-cloned-has-tests-directory-check
-    network_mode: host
-
-  github-done-collector:
-    build:
-      context: ./github-done-collector
-    network_mode: host
-
-  github-octo-repo-details:
-    build:
-      context: ./github-octo-repo-details
-    network_mode: host
-
-  github-octo-license-check:
-    build:
-      context: ./github-octo-license-check
-    network_mode: host
-
-  github-octo-languagues-check:
-    build:
-      context: ./github-octo-languagues-check
-    network_mode: host
-
