+++ scanners/github-cloned-repo-checks/src/get-dotignore-details.js
@@ -1,55 +1,7 @@
 // fs docs node https://nodejs.org/docs/v0.3.4/api/fs.html
-import * as fs from 'fs';
-import * as util from 'util';
-import { CheckOnClonedRepoStrategy } from './check-on-cloned-repo-strategy.js'
 
-function searchForFile(directory, targetFileName) {
-    // searches directory and returns array of file paths for any found targetFileNames
-    const files = fs.readdirSync(directory);
-    const foundFilePaths = []
-    for (const file of files) {
-        const fullPath = `${directory}/${file}`;
-        // Determine if it's a file or directory
-        const stat = fs.statSync(fullPath);
-        if (stat.isDirectory()) {
-            const subDirectoryPaths = searchForFile(fullPath, targetFileName);
-            foundFilePaths.push(...subDirectoryPaths); 
-        } else if (file.includes(targetFileName)) { //changing for part name too like dependabot.y
-        // } else if (file === targetFileName) {
-            foundFilePaths.push(fullPath); 
-        }
-    }
-    return(foundFilePaths)
-}
-
-async function searchFileForText(filePath, text) {
-    const readFileAsync = util.promisify(fs.readFile);
-    
-    const regEx = new RegExp(text);
-    const result = [];
-
-    try {
-        const contents = await readFileAsync(filePath, 'utf8');
-        let lines = contents.toString().split("\n");
-        lines.forEach(line => {
-            if (line && line.search(regEx) >= 0) {
-                // console.log('found in file ', filePath);
-                result.push(line);
-            }
-        });
-
-        if (result.length > 0) {
-            // console.log(text, "found text in file", result.length, "times");
-            return true;
-        } else {
-            // console.log("text not found in file");
-            return false;
-        }
-    } catch (error) {
-        console.error('An error occurred:', error);
-        throw error;
-    }
-}
+import { CheckOnClonedRepoInterface } from './check-on-cloned-repo-interface.js'
+import { searchForFile, hasTextInFile } from './searching-functions.js';
 
 
 export async function searchIgnoreFile(repoPath, ignoreFileName) {  //TODO - probably need to extract this again to test...
@@ -61,9 +13,9 @@ export async function searchIgnoreFile(repoPath, ignoreFileName) {  //TODO - pro
     if (ignoreFilePaths.length > 0) {
         for (const filePath of ignoreFilePaths) { //will this address null case?
             const ignoreFilePath = filePath.split('/').slice(3).join('/'); // removing ./temp-cloned-repo/${repo}
-            const hasDotenv = await searchFileForText(filePath, '.env') // temp value for logic  finding just .env in file (any form)
-            const hasDoubleStarSlashStarDotenv = await searchFileForText(filePath, '\\*\\*/\\*.env') // in this case is **/*.env
-            const hasDoubleStarSlashDotenvStar = await searchFileForText(filePath, '\\*\\*/.env.*') // looking for **/.env*
+            const hasDotenv = await hasTextInFile(filePath, '.env') // temp value for logic  finding just .env in file (any form)
+            const hasDoubleStarSlashStarDotenv = await hasTextInFile(filePath, '\\*\\*/\\*.env') // in this case is **/*.env
+            const hasDoubleStarSlashDotenvStar = await hasTextInFile(filePath, '\\*\\*/.env.*') // looking for **/.env*
             ignoreFileDetails.push ({
                 "ignoreFilePath": ignoreFilePath, 
                 "hasDotenv": hasDotenv, 
@@ -75,7 +27,7 @@ export async function searchIgnoreFile(repoPath, ignoreFileName) {  //TODO - pro
     }
 }
 
-export class DotGitIgnoreDetails extends CheckOnClonedRepoStrategy {
+export class DotGitIgnoreDetails extends CheckOnClonedRepoInterface {
     constructor(repoName, clonedRepoPath) { 
         super(repoName, clonedRepoPath); 
         this.clonedRepoPath = clonedRepoPath;
@@ -89,13 +41,9 @@ export class DotGitIgnoreDetails extends CheckOnClonedRepoStrategy {
             lastUpdated: Date.now()
         }
     }
-    checkName() {
-        this.checkName = 'gitignore'
-        return this.checkName
-    }
 }
 
-export class DotDockerIgnoreDetails extends CheckOnClonedRepoStrategy {
+export class DotDockerIgnoreDetails extends CheckOnClonedRepoInterface {
     constructor(repoName, clonedRepoPath) { 
         super(repoName, clonedRepoPath); 
         this.clonedRepoPath = clonedRepoPath;
@@ -109,8 +57,5 @@ export class DotDockerIgnoreDetails extends CheckOnClonedRepoStrategy {
             lastUpdated: Date.now()
         }
     }
-    checkName() {
-        this.checkName = 'dockerignore'
-        return this.checkName
-    }
+
 }
