+++ scanners/github-cloned-repo-checks/src/has-dependabot-yaml.js
@@ -1,35 +1,8 @@
-// fs docs node https://nodejs.org/docs/v0.3.4/api/fs.html
-import * as fs from 'fs'
-import * as util from 'util'
-import * as path from 'path'
-import { CheckOnClonedRepoStrategy } from './check-on-cloned-repo-strategy.js'
+import { CheckOnClonedRepoInterface } from './check-on-cloned-repo-interface.js'
 import { searchForFile } from './searching-functions.js'
-
-// api directory - libraries? - expand into endpoints etc...consume it?
-
-// export function searchForFile(directory, targetFileName) {
-//     // searches directory and returns array of file paths for any found targetFileNames
-//     const files = fs.readdirSync(directory);
-//     const foundFilePaths = []
-//     for (const file of files) {
-//         const fullPath = `${directory}/${file}`;
-//         // Determine if it's a file or directory
-//         const stat = fs.statSync(fullPath);
-//         if (stat.isDirectory()) {
-//             const subDirectoryPaths = searchForFile(fullPath, targetFileName);
-//             foundFilePaths.push(...subDirectoryPaths); 
-//         } else if (file.includes(targetFileName)) { //changing for part name too like dependabot.y
-//         // } else if (file === targetFileName) {
-//             foundFilePaths.push(fullPath); 
-//         }
-//     }
-//     return(foundFilePaths)
-// }
-
 
 export async function hasDependabotYaml(clonedRepoPath) {
     // searchForFile returns array of found file paths
-
     const dependabotFile = searchForFile(clonedRepoPath, "dependabot.y") // accounting for both .yaml and .yml
     if (dependabotFile.length > 0) {
         return true
@@ -38,8 +11,7 @@ export async function hasDependabotYaml(clonedRepoPath) {
     }
 } 
 
-
-export class HasDependabotYaml extends CheckOnClonedRepoStrategy {
+export class HasDependabotYaml extends CheckOnClonedRepoInterface {
     constructor(repoName, clonedRepoPath) { 
         super(repoName, clonedRepoPath); 
         this.clonedRepoPath = clonedRepoPath;
@@ -52,9 +24,5 @@ export class HasDependabotYaml extends CheckOnClonedRepoStrategy {
             metadata: null,
             lastUpdated: Date.now()
         }
-    }
-    checkName() {
-        this.checkName = 'hasDependabotYaml'
-        return this.checkName
     }
 }
