--- scanners/github-cloned-repo-checks/src/repo-check-context.js
@@ -1,7 +1,0 @@
-
-export class RepoChecker {
-    doRepoCheck(processor){
-        return processor.doRepoCheck()
-    }
-}
-
