--- scanners/github-save-to-database/.env.example
@@ -1,6 +1,0 @@
-DB_NAME=
-DB_URL=
-DB_USER=
-DB_PASS=
-NATS_URL=
-API_URL=
