--- scanners/github-save-to-database/Dockerfile
@@ -1,18 +1,0 @@
-FROM node:18-alpine
-
-ENV NODE_ENV production
-
-WORKDIR /app
-
-RUN apk --no-cache add curl
-
-COPY package*.json ./
-
-RUN npm ci
-
-# COPY . . 
-COPY ./.env ./.env.example ./
-COPY ./src ./src
-COPY ./index.js .
-
-CMD ["npm", "--no-update-notifier", "start"]
