--- scanners/github-save-to-database/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "github-save-to-database",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js","type": "module",
-  "scripts": {
-    "test": "NODE_OPTIONS=--experimental-vm-modules jest --coverage",
-    "lint": "eslint src",
-    "dbg": "node --experimental-vm-modules --inspect-brk node_modules/.bin/jest --runInBand --no-cache",
-    "start": "node --experimental-vm-modules index.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "arangojs": "^8.4.1",
-    "dotenv-safe": "^8.2.0",
-    "graphql-request": "^6.1.0",
-    "nats": "^2.16.0"
-  },
-  "devDependencies": {
-    "jest": "^29.6.4"
-  }
-}
